Kona: Gpu: Max GPU OPP to 944Mhz.
This works in tandem with Echelon Devfreq Gov. it dynamically changes frequencies,
but most important is that the top freq (944Mhz) acts as a burst frequency. unless user decides to use another gov
which causes that top freq to be utilized however that gov uses it. Echelon shouldnt "over" use the max frequency unless load is substantial.